LayoutTests/accessibility/mac/range-for-line-index-expected.txt

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS contenteditable.rangeForLine(0) is '{0, 15}'
PASS multilineContenteditable.rangeForLine(0) is '{0, 16}'
PASS multilineContenteditable.rangeForLine(1) is '{15, 1}'
PASS multilineContenteditable.rangeForLine(0) is '{0, 16}'
PASS multilineContenteditable.rangeForLine(1) is '{16, 6}'
PASS multilineContenteditable.rangeForLine(2) is '{22, 5}'
PASS textarea.rangeForLine(0) is '{0, 15}'
PASS multilineTextarea.rangeForLine(0) is '{0, 16}'
PASS multilineTextarea.rangeForLine(1) is '{16, 6}'
PASS multilineTextarea.rangeForLine(2) is '{22, 5}'
PASS textfield.rangeForLine(0) is '{0, 21}'
PASS successfullyParsed is true

TEST COMPLETE
this is a test.
this is a test.
this is a test.
hello
world
